Finaliza implementação do botão de identificação Termina a implementação do botão que identifica os equipamentos
